+++ dla3d/c/vecmath.h
@@ -15,19 +15,19 @@
  */
 #define KDT_DIM (3)
 typedef struct point {
-	float coord[KDT_DIM];
+	double coord[KDT_DIM];
 } point;
 
 typedef point vec;
 
-inline float vec_fmax(float f1, float f2);
-inline float vec_fmin(float f1, float f2);
-inline float vec_length(vec *a);
+inline double vec_fmax(double f1, double f2);
+inline double vec_fmin(double f1, double f2);
+inline double vec_length(vec *a);
 inline void vec_neg(vec *dst, vec *v);
 inline void vec_add(vec *dst, vec *a, vec *b);
 inline void vec_sub(vec *dst, vec *a, vec *b);
-inline float vec_dot_prod(vec *a, vec *b);
-inline void vec_scalar_mult(vec *dst, vec *a, float x);
+inline double vec_dot_prod(vec *a, vec *b);
+inline void vec_scalar_mult(vec *dst, vec *a, double x);
 inline void vec_copy(point *dst, point *src);
 inline void vec_zero(vec *dst);
 inline void vec_rand_unit(vec *dst);
